add delete inquiry

+++ routes/reservationsList.js
@@ -63,40 +63,44 @@ router.delete('/', function(req, res){
 		var reservationId;
 		var tripId;
 		
-		models.TRIP.findOne({
-			where: {
-				TRIP_ID: tripObj.dataValues.TRIP_ID
-			}
-		}).then(tripOb => {
-				var trip     = tripOb.dataValues;
-				
-				tripId       = trip.TRIP_ID
-				origin		 = trip.ORIGIN;
-				destination  = trip.DESTINATION;
-				trainId 	 = trip.TRAIN_ID;
-				date		  = trip.TRIP_DATE;
-				passengerId   = trip.PASSENGER_ID;
-				reservationId = trip.RESERVATION_ID;
-		});
-		
-		sequelize.query('call ADD_FREE_SEAT(?, ?, ?);', {
-				replacements: [origin, destination, trainId, date]
-		}).then(() => {
-			sequelize.query('SET FOREIGN_KEY_CHECKS = 0;').then(() => {
-				sequelize.query('DELETE FROM PASSENGER WHERE PASSENGER_ID=?;', {
-					replacements: [passengerId]
-				}).then(() => {
-					sequelize.query('DELETE FROM RESERVATION WHERE RESERVATION_ID=?;', {
-						replacements: [reservationId]
-					}).then(() => {
-						sequelize.query('DELETE FROM TRIP WHERE TRIP_ID=?;', {
-							replacements: [tripId]
-						}).then(() => {
-							sequelize.query('SET FOREIGN_KEY_CHECKS = 1;');
-						});
-					});
+		if(answer == 'yes'){
+				models.TRIP.findOne({
+					where: {
+						TRIP_ID: tripObj.dataValues.TRIP_ID
+					}
+				}).then(tripOb => {
+						var trip     = tripOb.dataValues;
+						
+						tripId       = trip.TRIP_ID
+						origin		 = trip.ORIGIN;
+						destination  = trip.DESTINATION;
+						trainId 	 = trip.TRAIN_ID;
+						date		  = trip.TRIP_DATE;
+						passengerId   = trip.PASSENGER_ID;
+						reservationId = trip.RESERVATION_ID;
 				});
-			});
-		});
+				
+						sequelize.query('call ADD_FREE_SEAT(?, ?, ?);', {
+								replacements: [origin, destination, trainId, date]
+						}).then(() => {
+							sequelize.query('SET FOREIGN_KEY_CHECKS = 0;').then(() => {
+								sequelize.query('DELETE FROM PASSENGER WHERE PASSENGER_ID=?;', {
+									replacements: [passengerId]
+								}).then(() => {
+									sequelize.query('DELETE FROM RESERVATION WHERE RESERVATION_ID=?;', {
+										replacements: [reservationId]
+									}).then(() => {
+										sequelize.query('DELETE FROM TRIP WHERE TRIP_ID=?;', {
+											replacements: [tripId]
+										}).then(() => {
+											sequelize.query('SET FOREIGN_KEY_CHECKS = 1;');
+										});
+									});
+								});
+							});
+						});
+			
+
+			}
 });
 module.exports = router;
